requirements.txt
# GUI Framework
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2

# Database
psycopg2-binary==2.9.9

# Data Processing
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
xlrd==2.0.1

# Template Engine
Jinja2==3.1.2
MarkupSafe==2.1.3
beautifulsoup4==4.12.2
lxml==4.9.3

# Email
email-validator==2.1.0

# Utilities
validators==0.20.0
python-dotenv==1.0.0
keyring==24.2.0
python-dateutil==2.8.2

# Build Tools
pyinstaller==5.13.0